add fuel-core-bin as default, add fuel-core-lib, add refactoring
