+++ Cargo.toml
@@ -12,7 +12,6 @@ members = [
     "packages/fuels-core",
     "packages/fuels-signers",
     "packages/fuels-test-helpers",
-    "packages/fuels-test-helpers-bin",
     "packages/fuels-types",
     "tools/fuels-abi-cli",
     "scripts/build-test-projects",
