+++ packages/fuels-abigen-macro/Cargo.toml
@@ -20,7 +20,7 @@ syn = "1.0.12"
 
 [dev-dependencies]
 ctor = " 0.1"
-fuel-core = { version = "0.8", default-features = false }
+#fuel-core = { version = "0.8", default-features = false }
 fuel-gql-client = { version = "0.8", default-features = false }
 fuels = { path = "../fuels", default-features = false }
 hex = { version = "0.4.3", default-features = false }
@@ -28,7 +28,3 @@ sha2 = "0.9.5"
 tokio = "1.15.0"
 tracing = {version = "0.1", default-features = false}
 tracing-subscriber = {version = "0.3", default-features = false, features = ["env-filter", "fmt"]}
-
-[features]
-fuel-core = ["fuels/fuel-core"]
-fuel-core-bin = ["fuels/fuel-core-bin"]
