+++ packages/fuels-test-helpers-bin/Cargo.toml
@@ -23,5 +23,6 @@ serde_with = "1.11"
 tempfile = "3.0.1"
 tokio = "1.15"
 
+anyhow = "1.0"
