+++ packages/fuels-test-helpers/Cargo.toml
@@ -9,12 +9,25 @@ repository = "https://github.com/FuelLabs/fuels-rs"
 description = "Fuel Rust SDK test helpers."
 
 [dependencies]
-fuel-core = { version = "0.8", default-features = false }
+fuel-core = { version = "0.8", default-features = false, optional = true}
+
 fuel-gql-client = { version = "0.8", default-features = false }
 fuels-core = { version = "0.14.1", path = "../fuels-core" }
-fuels-signers = { version = "0.14.1", path = "../fuels-signers", optional = true }
+fuels-signers = { version = "0.14.1", path = "../fuels-signers"}
 rand = { version = "0.8.4", default-features = false }
 tokio = "1.15"
 
+fuel-core-interfaces = { version = "0.8", default-features = false, optional = true}
+fuel-types = { version = "0.5", default-features = false, features = ["random"]}
+hex = { version = "0.4", features = ["serde"]}
+portpicker = "0.1.1"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = { version = "1.0", features = ["raw_value"]}
+serde_with = "1.11"
+tempfile = "3.0.1"
+
+
 [features]
-default = ["fuels-signers"]
+#default = ["fuels-signers"]
+fuel-core-bin = ["fuel-core-interfaces"]
+fuel-core-lib = ["fuel-core"]
