+++ packages/fuels-test-helpers/src/signers.rs
@@ -1,40 +1,47 @@
 use std::net::SocketAddr;
 
+#[cfg(feature = "fuel-core-lib")]
 use fuel_core::{model::Coin, service::Config};
+
 use fuel_gql_client::fuel_tx::UtxoId;
+
+#[cfg(not(feature = "fuel-core-lib"))]
+use fuel_core_interfaces::model::Coin;
 
 use fuels_signers::{provider::Provider, LocalWallet, Signer};
 
 use crate::{
-    setup_single_asset_coins, setup_test_client, wallets_config::WalletsConfig,
-    DEFAULT_COIN_AMOUNT, DEFAULT_NUM_COINS,
+    setup_single_asset_coins,
+    setup_test_client,
+    wallets_config::WalletsConfig,
+    // DEFAULT_COIN_AMOUNT, DEFAULT_NUM_COINS,
 };
 
-#[cfg(feature = "fuels-signers")]
+#[cfg(feature = "fuel-core-lib")]
 pub async fn launch_provider_and_get_single_wallet() -> LocalWallet {
     let mut wallets = launch_provider_and_get_wallets(WalletsConfig::new_single(None, None)).await;
 
     wallets.pop().unwrap()
 }
 
-#[cfg(feature = "fuels-signers")]
-pub async fn launch_custom_provider_and_get_single_wallet(node_config: Config) -> LocalWallet {
-    let mut wallet = LocalWallet::new_random(None);
+// #[cfg(feature = "fuel-core-lib")]
+// pub async fn launch_custom_provider_and_get_single_wallet(node_config: Config) -> LocalWallet {
+//     let mut wallet = LocalWallet::new_random(None);
+//
+//     let coins: Vec<(UtxoId, Coin)> = setup_single_asset_coins(
+//         wallet.address(),
+//         Default::default(),
+//         DEFAULT_NUM_COINS,
+//         DEFAULT_COIN_AMOUNT,
+//     );
+//
+//     let (provider, _) = setup_test_provider(coins, node_config).await;
+//
+//     wallet.set_provider(provider);
+//     wallet
+// }
 
-    let coins: Vec<(UtxoId, Coin)> = setup_single_asset_coins(
-        wallet.address(),
-        Default::default(),
-        DEFAULT_NUM_COINS,
-        DEFAULT_COIN_AMOUNT,
-    );
-
-    let (provider, _) = setup_test_provider(coins, node_config).await;
-
-    wallet.set_provider(provider);
-    wallet
-}
-
-#[cfg(feature = "fuels-signers")]
+#[cfg(feature = "fuel-core-lib")]
 pub async fn launch_provider_and_get_wallets(config: WalletsConfig) -> Vec<LocalWallet> {
     let mut wallets: Vec<LocalWallet> = (1..=config.num_wallets)
         .map(|_i| LocalWallet::new_random(None))
@@ -62,11 +69,53 @@ pub async fn launch_provider_and_get_wallets(config: WalletsConfig) -> Vec<Local
 
 // Setup a test provider with the given coins. We return the SocketAddr so the launched node
 // client can be connected to more easily (even though it is often ignored).
+#[cfg(feature = "fuel-core-lib")]
 pub async fn setup_test_provider(
     coins: Vec<(UtxoId, Coin)>,
     node_config: Config,
 ) -> (Provider, SocketAddr) {
     let (client, addr) = setup_test_client(coins, node_config).await;
+    (Provider::new(client), addr)
+}
+
+/// Default ----------------------------------------------------------------------------------------
+
+#[cfg(not(feature = "fuel-core-lib"))]
+pub async fn launch_provider_and_get_single_wallet() -> LocalWallet {
+    let mut wallets = launch_provider_and_get_wallets(WalletsConfig::new_single(None, None)).await;
+
+    wallets.pop().unwrap()
+}
+
+#[cfg(not(feature = "fuel-core-lib"))]
+pub async fn launch_provider_and_get_wallets(config: WalletsConfig) -> Vec<LocalWallet> {
+    let mut wallets: Vec<LocalWallet> = (1..=config.num_wallets)
+        .map(|_i| LocalWallet::new_random(None))
+        .collect();
+
+    let mut all_coins: Vec<(UtxoId, Coin)> = Vec::with_capacity(config.num_wallets as usize);
+    for wallet in &wallets {
+        let coins: Vec<(UtxoId, Coin)> = setup_single_asset_coins(
+            wallet.address(),
+            Default::default(),
+            config.coins_per_wallet,
+            config.coin_amount,
+        );
+        all_coins.extend(coins);
+    }
+
+    let (provider, _) = setup_test_provider(all_coins).await;
+
+    wallets
+        .iter_mut()
+        .for_each(|wallet| wallet.set_provider(provider.clone()));
+
+    wallets
+}
+
+#[cfg(not(feature = "fuel-core-lib"))]
+pub async fn setup_test_provider(coins: Vec<(UtxoId, Coin)>) -> (Provider, SocketAddr) {
+    let (client, addr) = setup_test_client(coins).await;
     (Provider::new(client), addr)
 }
